requirements.txt
asn1crypto==0.24.0
cffi==1.12.0
cryptography==2.5
dj-database-url==0.5.0
Django==2.1.7
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-rest-knox==3.6.0
djangorestframework==3.9.1
gunicorn==19.9.0
Pillow==5.4.1
psycopg2==2.7.7
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2018.9
six==1.12.0
whitenoise==4.1.2
